Cleaning up a bit.
