+++ metaconfigure/description.json
@@ -576,11 +576,12 @@
     "src/constants.hpp",
     "src/constants/CODATA2014.hpp",
     "src/constants/derived.hpp",
+    "src/constants/map.hpp",
     "src/constants/math.hpp",
     "src/constants/tags.hpp",
     "src/constants/derived/Insert.hpp",
-    "src/constants/derived/PlanckLength.hpp",
-    "src/constants/derived/ReducedPlanck.hpp"
+    "src/constants/derived/ReducedPlanck.hpp",
+    "src/constants/src/addUncertainty.hpp"
   ],
   "ignore pattern": "$^",
   "implementation files": [],
