+++ src/constants.hpp
@@ -14,6 +14,8 @@ namespace constants {
 #include "constants/math.hpp"
 #include "constants/tags.hpp"
 #include "constants/derived.hpp"
+#include "constants/map.hpp"
+#include "constants/src/addUncertainty.hpp"
 #include "constants/CODATA2014.hpp"
 
 } // namespace constants
