+++ src/constants/test/math.test.cpp
@@ -5,9 +5,9 @@
 
 SCENARIO( "Testing mathematical constants" ){
 
-  REQUIRE( 3.141592653589793 == njoy::constants::pi );
-  REQUIRE( 6.283185307179586 == njoy::constants::twopi );
-  REQUIRE( 2.718281828459045 == njoy::constants::e );
-  REQUIRE( 1.414213562373095 == njoy::constants::root2 );
+  CHECK( 3.141592653589793 == njoy::constants::pi );
+  CHECK( 6.283185307179586 == njoy::constants::twopi );
+  CHECK( 2.718281828459045 == njoy::constants::e );
+  CHECK( 1.414213562373095 == njoy::constants::root2 );
 
 }
